+++ client/domain/src/main/java/com/andone/memorip/domain/repository/PlaceRepository.kt
@@ -2,6 +2,7 @@ package com.andone.memorip.domain.repository
 
 import androidx.paging.PagingData
 import com.andone.memorip.domain.model.PlaceListItem
+import com.andone.memorip.domain.model.Region
 import com.andone.memorip.domain.model.request.PlaceCreateRequest
 import com.andone.memorip.domain.model.response.PlaceCreateResponse
 import com.andone.memorip.domain.model.response.PlaceDetailResponse
@@ -14,4 +15,5 @@ interface PlaceRepository {
     fun getPlaceList(): Flow<PagingData<PlaceListItem>>
     suspend fun uploadImage(file: File): Result<PlaceImageUploadResponse>
     suspend fun createPlace(place: PlaceCreateRequest): Result<PlaceCreateResponse>
+    fun loadRegions(): List<Region>
 }
